requirements.txt
# requirements.txt for DeepSeekMath-V2
# Python 3.10+ recommended

# ============================================================================
# Core Deep Learning Frameworks
# ============================================================================
torch>=2.1.0
transformers>=4.36.0
tokenizers>=0.15.0
accelerate>=0.25.0

# ============================================================================
# Reinforcement Learning
# ============================================================================
# Note: DeepSpeed will be added when GPU is available
# deepspeed>=0.12.0  # Uncomment when setting up GPU

# ============================================================================
# Data Processing & Scraping
# ============================================================================
datasets>=2.16.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0
html5lib>=1.1
selenium>=4.15.0  # For dynamic web scraping if needed
pandas>=2.1.0
numpy>=1.24.0

# ============================================================================
# Configuration & Experiment Management
# ============================================================================
pyyaml>=6.0.1
omegaconf>=2.3.0
hydra-core>=1.3.0
python-dotenv>=1.0.0

# ============================================================================
# Logging & Monitoring
# ============================================================================
tensorboard>=2.15.0
wandb>=0.16.0
tqdm>=4.66.0

# ============================================================================
# Evaluation & Metrics
# ============================================================================
scikit-learn>=1.3.0
scipy>=1.11.0

# ============================================================================
# Testing
# ============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# ============================================================================
# Code Quality
# ============================================================================
black>=23.12.0
flake8>=7.0.0
isort>=5.13.0
mypy>=1.7.0

# ============================================================================
# Utilities
# ============================================================================
typing-extensions>=4.9.0
pydantic>=2.5.0  # For data validation
rich>=13.7.0  # For better terminal output
click>=8.1.0  # For CLI tools

# ============================================================================
# Jupyter & Notebooks (Optional)
# ============================================================================
jupyter>=1.0.0
ipykernel>=6.27.0
ipywidgets>=8.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# ============================================================================
# Optional: Formal Math Tools (if you want to explore formal proofs later)
# ============================================================================
# sympy>=1.12  # Symbolic mathematics
# lean-client>=0.1.0  # Lean theorem prover interface (if available)